6.1.1 - исправления по замечаниям наставника
