+++ src/components/main/main.test.jsx
@@ -1,7 +1,7 @@
 import Main from "@components/main/main";
 import {Provider} from "react-redux";
 import configureStore from "redux-mock-store";
-import {ALL_GENRES} from "../../consts/consts";
+import {ALL_GENRES} from "@consts";
 import {testMockFilm} from "../../mocks/test-mock-film";
 
 const mockStore = configureStore([]);
